requirements.txt
Flask==3.1.2
openai>=1.0.0
google-search-results>=2.4.0
python-dotenv>=1.0.0
requests>=2.28.0
blinker>=1.9.0
itsdangerous>=2.2.0
Jinja2>=3.1.2
Werkzeug>=3.1.0
MarkupSafe>=3.0.0
